+++ NEWS.md
@@ -1,3 +1,17 @@
+# logLime 0.1.4 (7.02.2023)
+
+## New features
+
+-   `separate_logdata_types()`:
+    -   Handles data from the very, very first versions of the *logdataLimeSurvey* applet as well as data that do not contain information on input positions;
+    -   Returns additional variable describing problems in the data: `problemsNoActions` that marks respondent-screens for which there are no *actions* at all in the processed log-data;
+    -   Have new argument `inputsBoxCells` that enables to compute values that are further used to compute *relative* positions for table-format questions in a different way. In the new approach a box which upper left corner has relative coordinates (0,0) and bottom right corner has relative coordinates (1,1) is spanned not by the positions of inputs themselves but by boundaries (vertexes) of cells storing the inputs. This alternative specification is useful if one wants to analyse data in which different lengths of response scale were used.
+
+## Bug fixes
+
+-   `separate_logdata_types()` detects wider set of problems with time stamps.
+-   `compute_hovering()` constructs ids of objects it computes hovering for in a manner that do not cause problems if `target.id` is missing in the data.
+
 # logLime 0.1.3 (4.10.2022)
 
 ## Bug fixes
@@ -7,7 +21,7 @@
 ## Other changes
 
 -   `compute_relative_positions()` doesn't check any more whether data provided by the `actions` argument resembles `actions` element returned by `separate_logdata_types()` - now it requires it to contain only columns `pageX` and `pageY` (apart of those specified by the `screenId` argument).
--   Updats in README file.
+-   Updates in README file.
 
 # logLime 0.1.2 (10.06.2022)